sessionsPage add
